Changed 'set' table to 'exercise_set' in workout-scheduler sql file
While trying to use the sql-ts library to auto-generate typescript types
for the sqlite database I made, I found that the library would error out
when trying to parse the 'set' table in the workout-scheduler-v2.sql
file. Since 'set' is a keyword in sql, the error makes sense, so to
circumvent the problem, I renamed every instance of the 'set' table to
'exercise_set' in the sql file. While I haven't done so yet, I plan on
adjusting the visual database that we've been working on in dbdiagram to
reflect the table's name change.