preliminary auto-generated types produced for sqlite database
Right now, I've auto-generated a sqlite-types.ts file that stores all
the types of each table in the sqlite database. To make this work, I've
needed to install the ts-sql and sqlite libraries as dev dependencies.
You have to run 'node ts-sql.mjs' from the terminal to create the
sqlite-types.ts file. At this time, I also have to have an actual sqlite
database on hand to have the auto-generating script work (ts-sql, the
auto-generating library, needs an actual db and not just an sql script to
work). I plan on doing some work to skip having to use an actual db and try
to make the auto-gen script (ts-sql) just create an in-memory db and
discard it after it finishes auto-generating the types.
Since I'm not even sure if I will actually have a need for types with
regards to the sql tables, I'll need to make sure to delete these two
dependencies manually from the package.json should I choose to abandon
auto-generating table types.