Can auto-generate sql types w/ only sql script
In the previous commit, I had to use a pre-loaded sqlite database,
ddl.sqlite, in order to auto-generate db types. Now, I can auto-generate
these types by simply creating an in-memory database and deleting it's
(I think) backup file. This works and it's an improvement over having to
have an actual sqlite db in the project source code. One problem is that
I would have preffered not needing to delete the backup 'file' - I
couldn't figure out how to prevent it from being automatically made, but
I might be able to avoid that if I try out the better-sqlite library
instead of the current node-sqlite lib. We'll have to see about that
in another commit.