Set column properties for DB types to snake_case
Originally, I had columns to be in camelCase, but realized it would be a
pain in the butt when it comes time to assign my sql results to the
typed objects: the sql query results would be objects with the columns
in snake_case and then I'd have to assign those snake_case properties to
a typed object w/ camelCase properties. No thank you - that's why I just
decided to keep column properties in camel_case: that way object
assignment is much easier to do w/ far less typing.