Added all sql insert statements into pgdb-ddl.ts
pgdb-ddl.ts is meant to act as a module for holding an array of all the
create and insert statements for the sqlite db. Before this commit, I
only had the create statements array alongside 2 insert queries just so
I could grab some sample data. Now, all insert statements are also
available as an array of queries in the pgdb-ddl file.