Early proof of concept for making db 'globally' accessible
This is just super early work to prove to myself that you can pass a
database connection down to as a prop to child components and synchronize with it using React's useEffect function. Now I have a clearer
understanding of how I might take this proof of concept work and apply it to a
proper useContext function to make the sqlite database truly globally
accessible. React's useContext function basically does prop-drilling -
as in, it drills a given prop all the way down to every child in the
component tree. That's exactly what I need to make the db available to
every page. Again, this is early days. We'll see what we can accomplish
in a future commit.