Extracted Loading UI from child component to parent component
This is a small commit, but it something to keep note of: If you can
extract the loading screen stuff to a parent component, then that means
that the initial route (the tabs index in this case) can be our global
loading screen wherin we will wait for all resources such as our fonts,
auth session and namely, our database connect to finish loading. Then,
none our child components or other routes will have to deal with loading
these global reasources!