removed useStateHook magic type cast from useStorageState
I finally understand what the boolean for isLoading in the state array represents:
It only indicates whether we are loading in data for the FIRST TIME.
This means isLoading is only true on FIRST LOAD of a specifc key value and never again...
No wonder the loading screens I based off this variable never worked -
They would never render b/c isLoading was permanantly set to false after
the initial load of the session. Ultimately, this means that I will have
to rewrite this module or make a different loading state variable to handle making proper
loading logic.