upgraded expo to sdk 50
Had some warnings when upgrading, react-native-reanimated in particular
doesn't play nice with nativewind. I get a warning with reanimated
saying that the native version cannot be determined and the root cause
is the nativewind lib. Also, nativewind seems to be making it impossible
to use the app in the web browser. Nativewind is
nice, but it's causing a lot of pain to co-exist right now. I'll
consider disabling nativewind on web by editing the metro config later and
perhaps remove nativewind altogether, but that's for another commit.