changed _layout.tsx's initialRouteName from '/' to '(app)'
the initialRouteName variable is supposed to be used for indicating a
fallback/previous route to go back to when using a modal in a stack
layout. I made a mistake in an earlier commit changing the
initialRouteName to '/', but after reading the expo docs it seems that
the initialRouteName has to be the top-most group directory or file
(without the file-extension) within a layout to be valid. (app) fits as the top-most
group directory for acting as a fallback previous route when you need to
go back from a modal. Refer to the docs for more info: https://docs.expo.dev/router/advanced/router-settings/