updated all sql query code to use db context provider
As a side note, I deleted the initDB function in db-utils as it's no
longer neccessary w/ the context provider handling db initialization.
Furthermore, the deleteDB function in db-utils needs to be used w/
caution: Deleting the db won't trigger a re-render of the app, so if you
want to re-initialize the database after deleting it, you have to do a
full app refresh. Lastly, the 'hello' route no longer includes a button
to delete the db (it was just a placeholder button).