sql type interfaces use number for bigints
Bigints in an sqlite db are not exactly stored as bigints. I didn't
realize, but these bigints are treated as text by the interfaces so I
decided to remap bigints in the type definitions as numbers. This may
pose a problem in the future when we store truly big ints in the db, but
until then, I'm using Typescript's number type.