updated WorkoutCard Link to use href object
Instead of using a template string for the href that links to a workout
card based on id, I changed the href to be an object w/ a pathname and
query params. It's honestly just an aesthetic change and it's mostly
here to just let myself know links can be written in two ways.