removed footgun regarding 'active:' pseudo class
Nativewind has an 'active:' psuedo class modifier that lets you set
styling based on whether a component is pressed in. Problem w/ this is
that the 'active:' pseodo class was resetting the state of the title
input when trying to add a workout. I've also reported this in the
readme. Consider the 'active:' pseudo class modifier as unsafe and do
not use it. I will remove it from other files too at a later time.