early days SectionList implemented for workout page
I've also attempted to make the queries and data type safe, but I've run
into some issues w/ regards to react's SectionList sections prop which
doesn't seem to play nice with data of type 'UnifiedResistanceSet[] |
UnifiedCardioSet[]' It loooks like it only plays nice when the data
property of my section array is of type '(UnifiedResistanceSet |
UnifiedCardioSet)[]' This means react expects a unified array type, not
a split one. I'd like the split version as that's more the correct type
definition of my data, but we'll have to see about that in a later
commit.