less scuffed patch to appease SectionList data prop
The sectionData field where I grab all the exercises and sets into a
list of sections now returns the accurate type where it's guaranteed
that a Resistance exercise will only have data of UnifiedResistanceSet[]
and likewise for Cardio exercises. However, to appease the section prop
in SectionList, I've made a type cast in that specific spot that
appeases React's type requirements.