added dynamic details for # of workouts + calories
the # of workout and calories + average within the current graph view
are accurate however supporting the time row proved to be difficult. The
reason time wasn't added was due to the original postgres query not
being compatible with sqlite. the reason for this is due to the mode()
function not being supported in sqlite therefore a change and a custom
fixed is needed to circumvent the issue.