bug fixed: properly displays correct workout time
Issue with the incorrect amount of workout time was caused by the query
not accounting for rest time. having pulled the query from a larger
query this wasn't appearent and it may seem like the larger query has
this same issue. never the less rest_time is now accounted for.