made db reinitialization button on tab two
Previously, I had a 'Delete Database' button in the app for
reinitializing the database when I made changes to our sql startup
script, but that required us to manually reload the app to reload the
db. Now, I've updated the button's function to programatically reload
the app, making the button actually do it's job: deleting and
reinitializing the db.