bugfix: doesLocalDbExist input arg is used correctly
Turns out the function doesLocalDbExist wasn't using dbFilename, it's
input arg for checking if the dbFilename exists, but was rather just
using the defaultDatabase. This means we'd always check if
"next-sqlite.db" exists even if we entered a different file name. Now we
properly check for the file input given.