copy-pasta add-workout code to build-exercise route
Added code into build-exercise route that was mostly copy-pasta'd from
add-exercise in the workout route. Some stuff is stripped out from the
copy-pasta, but type definitions and the reducer are kept in tact. I'll have to
make sure to remove this code duplication in the future, but for now
it's fine.