renamed [exercise_class_id] route to build-exeercise
The reason for the renaming is that I found out how to add in query
params to a route. the build-exercise route now also takes in query
params that get the exercise_class_id, title, and exercise_type_id.
I initially used the [exercise_class_id] dynamic route scheme
b/c I thought that was the only way to add in params to get the
exercise_class info I needed for my exercise set building page, but now
that I know how to add in query params, I don't need the dynamic route
scheme for now.