create new workout modal that adds empty workout
The 'New Workout' button in Tab 2 now takes you to a modal
that simply asks for a new workout title with sensible defaults.
After pressing submit, you get redirected to the empty workout
to fill it in with new exercises as you please. The empty workout
page is not implemented yet, so that needs to be taken care of next.