made workout route context
Specifically, there is now a shared context within the workout route
that stores an array of the workouts and a workoutsDispatch function for
updating the workouts list. In this specific commit, the dispatch only
handles adding a new workout so that when the user adds a new workout in the
new-workout-modal, the workouts list gets updated immediately to reflect the
new addition - something we couldn't have if it weren't for the shared context.
Furthermore, the workouts querying functions were moved out to the _layout
route because that's where I define the context.