rename 'workout' route to 'workouts' (plural)
Using a plural 'workouts' route is the right choice because the 'workout' route
(now renamed 'workouts') is meant to represent ALL workouts available to the user.
Using a singular 'workout' route misrepresents its true meaning. Furthermore, I plan to
create a nested route inside the 'workouts' route in a later commit to represent a singular
workout, so it's even more important that we get the naming for routes corrected now.