added basic activity card template
activity card template is to be tailored to each gym session recorded by
the user in the app. the amount of cards is dependent on the time range
selected on the graph.
update to the template are to include actual titles, calories, dates and
svg image representing the type of workout engaged.