refactor: split workouts into multiple queries and removed workouts context
This commit does quite a lot of things. First, it splits
workouts data from tags data. Practically, I split up fetching
workouts and tags to make my data more "flat". The gaol of
flattening my data was so that mutations wouldn't trigger a full
refetch everything including workouts and tags. If we flatten the
data into separate lists (workouts, tags, and so on), if we mutate
one list, we don't have to refetch the others.

Furthermore, I completely removed the workouts context which I used for accessing
and mutating the workouts list from other components inside the workouts tab. Now
that I can fetch and cache workouts using React Query, there isn't really any need
for me to make a specific Context provider for it anymore. As a side-effect of
removing the context, I had to change up the New Workout Modal page since it relied
on the context to use React Query instead for fetching workout count and mutating
the workouts list.