expo v51 update
The expo upgrade to v51 introduced 2 problems that I had
to deal with:

1. With the update, the'useLocalSearchParams' function 
was typed to return a value or undefined. Since useLocalSearchParams
can now return potentially undefined data, I did a quick
fix where I do a 'value!' (non-null assert) on any values that I grab from
the useLocalSearchParams function. It's a bad, hacky fix,
but for now it's good enough to move forward quickly.

2. The update initially broke the build claiming "prop-types" was
not found in react. So, I found on stack overflow that install the
'prop-types' package from npm would fix the problem. I installed 'prop-types'
and it fixed the issue. If you see the prop-types package in the package.json
file in the future and wonder "Why does this exist?", now you know: it IS essential.