corrected sign-in and sign-out functions
sign-out/sign-in problem: whenever users wanted to sign out, they would be prompted
telling them that they will be redirected to be signed in when they're
actually signing out (this was a hack that worked for signing out). I've fixed this
hack so that you don't get any misinforming prompt. The fix is by using an
ephemeral session for the browser auth in the sign-in function (this option isn't
accounted for in the type system, but does actually work. Basically, be allowing an ephemeral session
the mobile browser never remembers the username and password which previously soft locked the user from
using a differnt user. See "https://github.com/expo/expo/pull/17792" for more info)

access token problem: the refresh token would get revoked instead of the access token.
now, we correctly revoke the access token. We'll worry about revoking the refresh token later.