update: delete and regenerate package-lock.json
There were some times when I rebased commits that generated
package-lock.json conflicts that I willy nilly merged and fixed.
Turns out, the right way to fix package-lock.json merge conflicts
is by going to the latest commit, running npm install, and overriding
the package-lock.json with whatever the npm install edited. So, to
start "fresh", I've just decided to delete and regenerate the package-lock.json
file. This caused a bug where nativewind was interfering with the <Pressable>
component styling using the pressed function args, so I regressed the nativewind
package to a version that didn't cause this issue.