feature: properly update UI when adding exercises
Previously, adding exercises wouldn't actually update the
exercise list UI when we redirected the user back to it. Now,
by utilizing react-query to invalidate the list UI's data when
adding an exercise to the db (handled by a useMutation), 
the list UI properly adds the new exercise as well. I also
added a utility type for making all params of an object
mandatary (not null and not undefined) to deal with sql
query types.