bugfix: fixed bug where adding new workouts would crash app
The bug was that after going through the add workout modal
and navigating to the empty workout page, turns out the
db query function I used didn't handle the case where the
exercise list was empty (as would be the case for a new workout)
and it just returned nothing (undefined). This was causing errors
and it's fixed to handle empty exercise lists. Furthermore, I
don't know how I missed it, but I guess I must have deleted
the workoutId param on my navigation code, so I just added
it back and things are peachy.