feature: made-bottom-sheet modal for deleting sets
Added a new dependency (bottom-sheet) to make bottom sheets
show up. Added a new react query mutation that deletes a set
given the exercise_set_id. Brought the two pieces together to
make a bottom sheet with a simple delete button that shows
up when you press the "more" icon on a set.