cleanup: removed training_cycle and training_days tables from db schema
While I wanted to have a concept of a training cycle - as in, the user
makes a schedule of workouts to do over a variable period of days. I've
come to realize that it's a bit too complicated for new users to grasp. It'd
be easier for users to just schedule a workout on a specific date and display
that schedule as a list. Maybe then users could have the option to repeat that schedule.
It's just ideas for now, but scheduling a workout via training_cyles and whatnot
is out of the picture.