db refactor: tied exercise_class to exercise_equipment in db schema
Initially, I had the exercise table (representing an instance of an exercise_class)
tied to a piece of eqipment, but I thought that was weird. It implied you could have
2 bench press exercises, one using dumbbells and another using a barbell (and maybe even
a 3rd using a machine - aka chest press). Thing is, idk if users want separate exercise
instances or just have one exercise instance that lets them change what equipment they use
on the fly. I liked the latter option. So, even though I now have equipment_ids tied to
exercise_classes instead of instances, the choice of equipment isn't really a hard and fast
rule. You choose the equipment for an exercise_class, but it's only the DEFAULT choice, after
which you can choose to change equipment. And the only real difference a change in equipment
makes is that we should adjust the weights of the new equipment to match as close as possible
to the target weight of the default eqipment. So, if you had a barbell bench press, then switched
to dumbbells, we'd halve the weights to 60lbs (display weight per dumbell) if you're using fixed weight dumbbells
and we'd get the closest dumbbell weight we could to 60lbs if you were planning on using adjustable weight dumbells.