refactor: moved workout and exercise queries/mutations to custom hooks
The plan is to colocate db queries and react query/mutations into
separate hook files that are specific to each use-case. For example,
if I want get the list of workouts, I calle the 'useWorkouts'
hook. With the refactor, this hooks logic and db query functions
are all colocated inside of a single file: 'workout-ids.ts'. This
way all the functionality for getting a workout list lives in one
colocated file.