feature: able to add custom exercise
For UI, I added a custom exercise card component that the user fills out.
The UI uses a new package for implementing dropdown menus.
For db/react-query logic, I modified the db schema to have a body_part column
in the exercise_class table and I made a react-query hook for adding a custom
exercise.