bugfix: fixed foreign key constraints (and cascading deletes) not working after app reloads
This bug manifested only after app reloads. When a user deleted the most recently added exercise
in the app and then added a new one, the new exercise would retain all sets of the previous exercise.
Why did this happen? Turns out, it has to do with the `PRAGMA foreign_keys = ON` constraint in sqlite.
I put this constraint in the sqlite init script, but it doesn't run when the app is reloaded in dev mode.
This happens because the foreign key constraint instruction has to be run on every new connection. When
you reload the app, you lose your db connection and have to make a new one. I explicitly wrote in the db init function
that the db script shouldn't run if the db file already exists and since I only put the constraint in the init script,
any time I reloaded the app, the foreign key constraint no longer got enabled.
I incorrectly thought that the `PRAGMA foreign_keys = ON` constraint would still be active given that I initialized the
database with it on in the init script, but it turns out the foreign key constraint is held on a per-connection level.
As in, the constraint isn't something you set in stone on db creation, it's something you have to set when you establish a
db connection. The way I corrected this bug was by making sure to run the `PRAGMA foreign_keys = ON` instruction on every render
of the db context holder in the app and there you go, bug fixed.