made derived state for active sets of a given exercise
Problem: We want to get all the sets for a given exercise, but we don't store
sets per exercise, we just store all of them in a normalized table. So, we get
that sets "table" with `useActiveWorkoutSetEntities` within an ActiveExerciseCard
component and grab the specific sets from there. We wrap this grab operation in a
useMemo to prevent re-computing if the setIds list for that exercise didn't change.