performance improvement: only 1 card re-renders when adding a set to an active exercise
I accomplished this by using a custom equality function in the selector for active sets.
This may introduce a problem when trying to re-order sets as the equality fn checks
if the lengths of the current setIds array and updated setIds array match. Since reordering won't
change the lengths, we'll have to change the equality fn at some point if we add a reordering
feature for active sets.