made active sets use set ids instead of set objects
Instead of passing an ActiveSet object as a prop to a set component,
I pass the setId to the component instead and the compoent can grab
the set it cares about from there. This makes it so that I no longer
have to do the set entities "hack" that I used earlier where I had to do
'useActiveWorkoutSetEntitiesByIds' with the incorrect, but useful at the
time, equality check. This commit avoids all that roundabout hacking.