2 fixes: added sets are always unchecked & first set bug fixed
By "first set bug fixed" I'm talking about the fact that the first set
of the first exercise wasn't being used as the next set placeholder (you
know if the last set of an exercise is 3min rest, 12 reps, 90kg, we make
it so that the next added set is also the exact same). This behavior was
not being followed in the case where you start the active workout with a
single set for the first exercise. To sum up why this happened it's
because the number 0 is falsey. Look at the diff to this commit and
you'll get what I mean.