add exercise name to active exercise UI + bugfix
Ended up fixing a silent bug where exercise_class_id was not being added
to the ResistanceSection query. So, the exercise_class_id property was
actually undefined this whole time which meant things were not as we
expected. The data layer mismatch strikes again. Anyways, we can now
properly get the exercise class name for each exercise card in the
active workout screen.