fix theme provider to use custom theme rather than defaults
This was a long running issue that I had with the app. All the
backgrounds and tab bars and tab headers would all basically not follow
the colors that I had setup in my custom color scheme and the reason was
because the app using this `ThemeProvder` component that was being
supplied default theme configs. So, I changed the theme configs to use
my own custom themes and things are fixed. Root/parent views no longer
have a differnt background color than my custom components. The
background colors all match up now. However, there's lots of pages and
screens in the app that are currently using static colors, so I'll need
to go in and make them adapt based on the color scheme in another
commit. Have fun future me!