got ui sliding loader and timer bugs fixed!
The bugs had to do with the timer not updating to the correct rest
values when the app is in the background or inactive. That's fixed, the
timer updates correctly. Another bug had to do w/ the slider ui (the
blue sliding window inside the rest cell that activates when you finish
a set). This slider wouldn't be at the correct position when you
navigated away from the page or put the app in the background, but now
that too is fixed. So, all good there!