refactored Workout MiniPlayer to navigate - not push - to new route
The purpose of this change is that `push` in `router.push` caused a bug
where if the user pressed the mini player button in rapid succession,
they would end up navigating to the active workout multiple times. For
example, 3 quick taps would "push" the active workout screen out 3
times, so if you went back from one screen, you'd return to a duplicate
screen 2 more times. Using `router.navigate` instead prevents this bug
as it only pushes a new route if it doesn't already on the navigation
stack.