made plate heights scale w/ weight
added a function to make it so that plate heights would scale
with weight. It linearly scales from a set max height and a set min
height making it impossible to have plates that are too big that they
clip or plates too small that they can't be seen. Damn, I'm good, (I
should also mention it's nothing but basic algebra, but I must gloat to
the void at least)