made button-press visual effects for keypad buttons (WeightAdjustView)
As in, I mimicked TouchableHighlight. I made an underlay view that
is white by default and when you press the button, the opacity of the
button reduces revealing the white underlay beneath. The only downside
of this is that the overlay and underlay effectively have to have
different styles that potentially get duplicated (i.e if overlay has
rounded borders, you have to also style the underlay with the same
rounded borders, you don't get any kind of smart copying behavior to
make sure both overlays match their shapes). So, be careful with this
KeypadButton component: you have to style things properly or the
underlay beneath will clip through the overlay.