made new colors to replace old ones + codegen
Made a whole new color pallete based on oklch color space
for making colors that have percuptually uniform brightness.
As in, 10% lightness colors are 10% lighter to the human eye.
Made the new color pallete alongside a codegen script to make
the colors into an object. Made the codegen script in case I want to
make derivative colors off the base color pallete. I don't use any
of the colors in this commit, but will do a refactor to use them later.