transition to drizzle for all db work
In this commit, I've installed drizzle an ORM framework for making
type-safe queries to a database. The code changes can be described in
3 parts: database schema generation scripts, db migration scripts, and
drizzle integration on the front-end. I opted to generate drizzle's
required `schema.ts` by reading from a sample database that I also
auto-generate with a script. That's the first 2 parts in a nutshell. The last is just
using drizzle: I wrote a simple `useDrizzle` hook to easily access the
drizzle client and write a simple test query to try it out. It may look
like a lot of changes, but it's mostly just integration work.