removed days_of_week & workout_days from sql schema
I did this b/c days_of_week is utterly useless and workout_days is a
feature I'm not sure when I'm actually going to implement. The idea
behind workout_days was to indicate what days out of the week you'd
schedule your specific workouts, but honestly, I'll worry about that
feature when I get to it and maybe then I'll modify the schema again to
incorporate workout_days, but until then I'm ditching it.