tightened up one-to-one relationships for exercise_set table
To elaborate, I modified the table relationships slightly for the
exercise_set and resistance_set/cardio_set tables in the SQLite
database. The relationship was implicitly one-to-one, as in, a
resistance_set/carido_set was associated with just one exercise_set, but
now I've gauranteed that by making the resistance_set/cardio_set tables'
primary key also act as a foreign key to the exercise_set table. This
ensures that every exercise_set row has exactly one associated
resistance_set/cardio_set, but not both.