fixed navigation bug for add-exercise page
I fixed a bug where after adding an exercise in a workout template, the
onPress callback of an exercise would navigate to a duplicate page of
the workout making you have to go back twice to get back to the original
workout page. Now, when you press an exercise on the add-exercise menu,
it correctly dismisses the menu and returns to the workout page.