integrated drizzle into app with react context
Here's the problem this commit solves: you normally get access to
drizzle queries by a 2-phase process: you call `const db = useSQLiteContext()` and
then pass the db from that into `const drizzleDb = drizzle(db)` and THEN
you can use a drizzle query. This is cumbersome and also involves a
memory penalty as every `drizzle(db)` call creates a separate reference
to the same underlying database. To solve this problem of constantly
making new references, I created a react context that instantiates a
single reference to the database with drizzle and provides access
through a `useDrizzle` hook. This way, we instantiate a drizzle db
instance once and always return that same instance when we need it in a
data fetching component.