remove unused drizzle database type from db-utils
I added this type way earlier on, but it was a mistake so I'm just
fixing it by removing it from here. I'm going to use this data type
later for being able to add the drizzle database handle as a parameter
to query functions later on though, so the type will be used but in a
different file.