cordoning off graph page while fixing issues regarding db handle
looks like reloading the app as of right now causes it to crash when
loading back into the main graph page. Think it has to do with trying to
access data with a database handle that's already closed somehow. Need
to look into it later, but I'm just cordoning this whole page off for
now until I get to revamping it.