minor refactoring to active workout page
only notable thing is after updating Typescript with the upgrade to Expo
SDK 53, the `setInterval` returns a `number` type instead of
`NodeJSTimeout` so I just changed the type and we're all set there. Also
renamed the set and exercise ID incrementing functions to indicate more
clearly that we're not incrementing a timer, but rather the id of the
set or exercise as we add more sets and exercises to our active workout.