removed quickfix for swipeables throwing reanimated warnings
The quickfix I'm talking about refers to telling expo to explicitly not
use it's reccomended version of react-native-gesture-handler and
explicitly set a more updated version of it. The more updated version,
at least anything from version 2.21.0 (of react-native-gesture-handler) onward fixed the swipeables
warnings for react-native-reanimated swipeables. With Expo SDK 53
installed, the latest react-native-gesture-handler version it supports
is v2.24.0, so the issue is fixed without the patch anymore, hence why
I'm removing it.