fixed error when active set weight is lower than 2.5kg
Had an issue where calculating the min and max plate weights for the
plate inventory would lead to undefined values when the weight is too
low (lower than the lowest plate weight of 2.5). This would return an
empty plate configuration list and I'd try to access the weight of the
largest and smallest weight in said list. Since the list is empty in
this case, accessing the wieght property would lead to an error "trying
to access weight poroperty of undefined". Fixed the issue by using the
`at` function which will return undefined instead of throw an error if
the plate configuration is empty due to too low a weight.