first phase of graph prototyping complete
Graph work isn't actually done, but for the prototyping stage, we've
come along pretty far. We have smooth scroll/scroll-off. We have a
segmented time range slider. However, we still need to make it so that
dates are shown instead of numbers on the x-axis. My next plan is to
make randomized workout session data so that we can better test the
graph's time segmenting feature and make sure each time range properly
works.